+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added Bing news searcher as class `BingNews`.
+
 ## [1.2.2] - 2021-11-20
 
 - SharpX upgraded to version 1.0.3.